refactor: ArgumentResolver 로직 변경
