+++ backend/src/main/java/com/stampcrush/backend/config/resolver/CustomerAuthArgumentResolver.java
@@ -1,0 +1,34 @@
+package com.stampcrush.backend.config.resolver;
+
+import com.stampcrush.backend.auth.application.util.AuthTokensGenerator;
+import com.stampcrush.backend.auth.application.util.BearerParser;
+import lombok.RequiredArgsConstructor;
+import org.springframework.core.MethodParameter;
+import org.springframework.http.HttpHeaders;
+import org.springframework.stereotype.Component;
+import org.springframework.web.bind.support.WebDataBinderFactory;
+import org.springframework.web.context.request.NativeWebRequest;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.method.support.ModelAndViewContainer;
+
+@RequiredArgsConstructor
+@Component
+public class CustomerAuthArgumentResolver implements HandlerMethodArgumentResolver {
+
+    private final AuthTokensGenerator authTokensGenerator;
+
+    @Override
+    public boolean supportsParameter(MethodParameter parameter) {
+        return parameter.getParameterType().equals(CustomerAuth.class);
+    }
+
+    @Override
+    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
+        String authorization = webRequest.getHeader(HttpHeaders.AUTHORIZATION);
+
+        String accessToken = BearerParser.parseAuthorization(authorization);
+        Long customerId = authTokensGenerator.extractMemberId(accessToken);
+
+        return new CustomerAuth(customerId);
+    }
+}
